X.26 (#1554)
* Initial branch of x.26

* Review deprecated methods, change build to error on deprecation (#1551)

* Review deprecated methods, change build to error on deprecation

* <release>8</release>

* Fix compilation on Java 21

* Don't use OS.getUsrDir() yet

* Use release 8 only for Java 11+

* Turn off implicit annotation processing

* Turn off implicit annotation processing

* Add expected exception

* Add ignored error

* Use a common configuration for the compiler-plugin where deprecation … (#1553)

* Use a common configuration for the compiler-plugin where deprecation warnings cause the build to fail

* Use a common configuration for the compiler-plugin where deprecation warnings cause the build to fail

* Deprecate RollCycles legacy code.

* Deprecate due x.26.

* Tidy up compiler setting and some warnings

* Reduce the number of raw types

* remove deprecated MultiQueue classes

* Remove deprecated methods (#1557)

* Remove deprecated methods

* remove deprecated SuppressWarnings

* resolve Javadoc issues

* Added fail on warning lint for Javadoc, and rawtypes fixes for Bytes (#1558)

* import net.openhft.chronicle.wire.Sequence

* serializable warning suppressed

* Fix warnings following rebase with develop

* Add methods for clear() and emptyHistory() to reduce direct references to VanillaMessageHistory

* Fix warnings following rebase with develop

* PretouchUtil deprecated

* PretouchUtil deprecated

* Fix deprecation issues in Java 17+

* Fix newlines

* redundant newline

* redundant spaces

* </p> tag in javadoc not allowed

* Javadoc fixes

* Javadoc fixes

* TimeoutPauser -> YieldingPauser

* ChronicleQueue#createPretoucherEventHandler

* ChronicleReader suppress empty message. Closes #1579

* TimeoutPauser -> YieldingPauser

* parent poms -> x.26.0

* Where marked for removal, also mark for deprecation

---------

+++ src/main/java/net/openhft/chronicle/queue/internal/main/InternalBenchmarkMain.java
@@ -136,7 +136,7 @@ public class InternalBenchmarkMain {
                 System.out.println(sb);
             }
 
-            next += messageSize * 1e9 / (throughput * 1e6);
+            next += (long) (messageSize * 1e9 / (throughput * 1e6));
             long delay = next - System.nanoTime();
             if (delay > 0)
                 LockSupport.parkNanos(delay);
